requirements.txt
# Core dependencies
python-dotenv==1.1.1
pydantic==2.11.7

# Natural language processing
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
openai==1.95.1

# Document processing
pypdf==5.7.0
faiss-cpu==1.11.0

# Database and storage
sqlite3_api==2.0.4

# Agent tools
requests==2.32.4

# UI Formatting
tabulate==0.9.0
